requirements.txt
pdfminer.six==20201018
pandas~=1.1.5
